User Registration
2:41:45 - 2:49:00 Django Tutorial

+++ Django Tutorial/studybud/base/views.py
@@ -5,6 +5,9 @@ from .forms import RoomForm
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse, HttpResponseGone
+
 
 # Create your views here.
 
@@ -23,6 +26,10 @@ from django.contrib import messages
 #     },
 # ]
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == "POST":
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -63,21 +70,27 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
             form.save()
             return redirect('home')
+
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!!")
+
     context = {'form': form}
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -86,8 +99,11 @@ def updateRoom(request, pk):
             return redirect('home')
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You're not allowed here!!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
